Refactor Choice into its own component.
